+++ Anti  Spoofing  Project  using  DL/Anti_Spoofing_Project_using_DL/Dataset/live_selfie/0001e96803--6239bb2420237d576aade0ea__Moto G30.jpg
