+++ Anti  Spoofing  Project  using  DL/Anti_Spoofing_Project_using_DL/WebApp/README.md
@@ -1,0 +1,110 @@
+# **🔴 Project Title: Anti-Spoofing Project**
+
+# **🔴 Goal:**
+The main goal of this project is to create a deep learning model that can accurately identify real images from various types of spoofed images, including replays and cut-outs. The purpose is to develop an effective anti-spoofing system for enhanced security.
+
+# **🔴 Dataset:**
+The dataset used for this project can be found at: [Dataset Link](https://www.kaggle.com/datasets/tapakah68/anti-spoofing). It contains a collection of real and spoofed images, which will be used for training and evaluation.
+
+# **🔴 Description:**
+This project focuses on implementing a deep learning model for anti-spoofing. The goal is to accurately differentiate between images of different kind such as real image, replay, prinout, layout and cut out. to enhance security measures. Various algorithms and models will be explored to achieve the desired results.
+
+# **🔴 What I Have Done:**
+To develop the anti-spoofing Web App, the following steps were followed:
+
+Folder Structure:
+Anti  Spoofing  Project  using  DL
+ |-Anti_Spoofing_Project_using_DL
+    |- Dataset
+      |- cut-out printouts
+      |- live_selfie
+      |- live_video
+      |- printouts
+      |- replay
+      |- README.md
+    |- Images
+      |- Image files
+    |- Model
+      |- anti_spoofing_project_using_dl.ipynb
+      |- siamese_model.h5
+      |- README.md
+    |- Web App
+      |- app.py
+      |- templates
+      |- static
+      |-Uploads
+      |- Anti_Spoofing_final_demo.mp4
+      |- README.md
+    |- requirements.txt
+
+1. Setting Up the Environment
+- Backend (Flask) :
+Install Flask and dependencies: Use pip or pipenv to install Flask and any other necessary Python libraries (e.g., OpenCV, TensorFlow).
+
+2. Loading and Preparing the Model
+- Load the pre-trained model: Use TensorFlow/Keras to load the trained anti-spoofing model.
+- Define helper functions: Create functions to process input images/videos and make predictions using the model.
+
+3. Developing the Backend
+-Flask Routes
+- Upload route: Serve the main HTML page.
+- Process route: Handle file uploads from the user (images or videos).
+- Prediction route: Process the uploaded files, run them through the model, and return the results.
+
+4. Building the Frontend
+- HTML Templates
+- Main Page: Include a form for uploading images/videos and a section to display results.
+-Result Display: Use placeholders in the HTML to dynamically show the prediction results.
+- CSS Styling :
+Styling Elements: Style buttons, forms, and results display for better user experience.
+
+5. Integrating Backend and Frontend
+Form Handling: Ensure the form in the HTML template correctly posts data to the Flask backend.
+
+6. Testing 
+Local Testing: Run the Flask app locally to test all functionalities and ensure the model predictions are correct.
+Debugging: Checked for and fixed arising issues or bugs.
+
+# **🔴 Models Used:**
+The following models were used in this project:
+
+1. Convolutional Neural Network (CNN):
+   - Used for image classification tasks.
+   - Provides a baseline for comparison.
+
+2. Recurrent Neural Network (RNN):
+   - Used for sequence-based analysis.
+   - Explored for its potential in anti-spoofing tasks.
+
+3. Siamese Neural Network:
+   - Chosen over CNN and RNN models.
+   - Specifically designed for image similarity and one-shot learning tasks.
+   - Capable of accurately measuring similarity between images.
+   - Well-suited for detecting and classifying between images.
+
+The Siamese model was selected for its unique architecture and its ability to accurately compare and measure similarity between images, making it a suitable choice for the anti-spoofing task.
+
+# **🔴 Libraries Needed:**
+The following libraries are required for this project:
+
+-Flask
+- TensorFlow
+- Keras
+- OpenCV
+- NumPy
+- Matplotlib
+-Werkzeug.utils
+-Scikit-learn
+- Pandas
+-Os
+
+# **🔴 Conclusion:**
+In conclusion, the Siamese model outperformed the CNN and RNN models in terms of accuracy for the anti-spoofing task. Its unique architecture, specifically designed for image similarity and one-shot learning, proved to be effective in accurately differentiating between different types of images. The Siamese model provides a robust solution for enhancing security measures against various types of image-based spoofing attacks.
+
+# **Author:**
+
+Pabitra Kumar Bera
+
+GitHub: https://github.com/PABITRA34
+
+LinkedIn: https://www.linkedin.com/in/pabitra34/
