+++ Anti  Spoofing  Project  using  DL/Anti_Spoofing_Project_using_DL/WebApp/templates/upload.html
@@ -1,0 +1,38 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Upload Status</title>
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css">
+    <style>
+        body {
+            background-image: url('https://www.pymnts.com/wp-content/uploads/2019/12/Biometric-Authentication-Digital-ID-Spoofing.jpg');
+            background-size: cover;
+            background-repeat: no-repeat;
+            background-position: center;
+            height: 100vh;
+        }
+    </style>
+</head>
+<body>
+    <div class="container mt-5">
+        <div class="card">
+            <div class="card-body">
+                {% if message %}
+                    <div class="alert alert-success" role="alert">
+                        {{ message }}
+                    </div>
+                {% elif error %}
+                    <div class="alert alert-danger" role="alert">
+                        {{ error }}
+                    </div>
+                {% endif %}
+                <a href="/process" class="btn btn-primary mt-3">Detect Spoofness</a>
+                <a href="/" class="btn btn-secondary mt-3">Go Back</a>
+            </div>
+        </div>
+    </div>
+</body>
+</html>
+
