+++ Anti  Spoofing  Project  using  DL/Anti_Spoofing_Project_using_DL/requirements.txt
@@ -1,0 +1,8 @@
+Flask
+opencv-python
+numpy
+werkzeug.utils
+scikit-learn
+tensorflow
+keras
+os
