--- Anti Spoofing Project using DL/WebApp/readme.md
@@ -1,21 +1,0 @@
-## Web Application for Brain Tumor Detection
-
-### Goal 🎯
-The main goal of is to provide a user-friendly web application for the detection of brain tumors using deep learning models. The purpose of the application is to assist healthcare professionals in accurately diagnosing brain tumors in patients.
-
-### Model(s) used for the Web App 🧮
-The web application utilizes three main models for brain tumor detection:
-1. **VGG16 Transfer Learning Model**: A pre-trained model on ImageNet, used for its superior performance in image classification tasks.
-2. **Custom Convolutional Neural Network (CNN)**: Specifically designed for brain tumor detection, trained on a dataset of brain MRI images.
-3. **Multilayer Perceptron (MLP) Based on TensorFlow**: An MLP architecture implemented using TensorFlow, serving as a simpler model for comparison.
-
-### Video Demonstration 🎥
-[![](https://github.com/TheNaiveSamosa/DL-Simplified/blob/700469f82238ca2579e09a5a02c5afb69cfb327c/Brain%20Tumor%20Detection/Web%20App/Screenshot%202024-05-13%20221628.png)](https://github.com/TheNaiveSamosa/DL-Simplified/assets/112872086/0979d5eb-5f8f-4e5c-8c64-df81ff0ab470)
-
-
-### Signature ✒️
-Aditya Khamitkar
-[![Twitter](https://img.shields.io/badge/Twitter-%40Couch_Potatoh_-blue?style=flat&logo=twitter)](https://twitter.com/Couch_Potatoh_)
-[![LinkedIn](https://img.shields.io/badge/LinkedIn-Aditya_Khamitkar-blue?style=flat&logo=linkedin)](https://www.linkedin.com/in/adityakhamitkar/)
-[![Instagram](https://img.shields.io/badge/Instagram-couch_potatoh_-blue?style=flat&logo=instagram)](https://www.instagram.com/couch_potatoh_/)
-[![Reddit](https://img.shields.io/badge/Reddit-The_Cactus_Flower-blue?style=flat&logo=reddit)](https://www.reddit.com/user/The-Cactus-Flower/)
