--- Anti Spoofing Project using DL/requirements.txt
@@ -1,5 +1,0 @@
-opencv-python
-numpy
-scikit-learn
-tensorflow
-keras
