--- Anti_Spoofing_Project_using_DL/Dataset/anti-spoofing.csv
@@ -1,10 +1,0 @@
-live_selfie,live_video,cut-out_printouts,printouts,replay
-live_selfie/0001e96803--62294efac72a824e5bb369f3__Xiaomi note 10s.mp4,live_video/0001e96803--62294efac72a824e5bb369f3__Xiaomi note 10s.mp4,cut-out printouts/0001e96803--62294efac72a824e5bb369f3.mp4,printouts/0001e96803--62294efac72a824e5bb369f3.mp4,replay/0001ffba3c--62880080572a894c5df8f427__Samsung Galaxy J2 Core __Samsung Galaxy A22.mp4
-live_selfie/0001e96803--6228928cb608822f9f53589e__Xiaomi redmi note 9 pro.mp4,live_video/0001e96803--6228928cb608822f9f53589e__Xiaomi redmi note 9 pro.mp4,cut-out printouts/0001e96803--6228928cb608822f9f53589e.mp4,printouts/0001e96803--6228928cb608822f9f53589e.mp4,replay/0001ffba3c--6289196875a5cf63cd0cb302__Samsung galaxy s20 __IPhone 6s plus.mp4
-live_selfie/0001e96803--62457bc94ff454214e8af551__Infinix note 11.mp4,live_video/0001e96803--62457bc94ff454214e8af551__Infinix note 11.mp4,cut-out printouts/0001e96803--62457bc94ff454214e8af551.mp4,printouts/0001e96803--62457bc94ff454214e8af551.mp4,replay/0001ffba3c--628a68b5180bc205cf8704d2__Vivo y10__Honor 10 lite.mp4
-live_selfie/0001e96803--6239bb2420237d576aade0ea__Moto G30.mp4,live_video/0001e96803--6239bb2420237d576aade0ea__Moto G30.mp4,cut-out printouts/0001e96803--6239bb2420237d576aade0ea.mp4,printouts/0001e96803--6239bb2420237d576aade0ea.mp4,replay/0001ffba3c--628c9dbb2579312f1ac7ee6f__С‚РҐРё+РЅС‚РҐРёС‚Р¦РЎС‚РҐРёС‚РҐР­С‚РҐРґ+-С‚РҐРґ+РІС‚РҐРёС‚РҐР¬С‚РҐРёС‚Р¤Р’  С‚РҐРё+Р™20__С‚РҐРё+РЅС‚РҐРёС‚Р¦РЎС‚РҐРёС‚РҐР­С‚РҐРґ+-С‚РҐРґ+РІС‚РҐРёС‚РҐР¬С‚РҐРёС‚Р¤Р’ С‚РҐРёС‚Р¦РЎ22.mp4
-live_selfie/0001e96803--62663cf6d21d14509973bcb4__Poco x3.mp4,live_video/0001e96803--62663cf6d21d14509973bcb4__Poco x3.mp4,cut-out printouts/0001e96803--62663cf6d21d14509973bcb4.mp4,printouts/0001e96803--62663cf6d21d14509973bcb4.mp4,replay/0001ffba3c--628e621c6789b1401e3cb184__Honor 50__Samsung note 9.mp4
-live_selfie/0001e96803--623955e76e1ee5175569a12b__HONOR 10I .mp4,live_video/0001e96803--623955e76e1ee5175569a12b__HONOR 10I .mp4,cut-out printouts/0001e96803--623955e76e1ee5175569a12b.mp4,printouts/0001e96803--623955e76e1ee5175569a12b.mp4,replay/0001ffba3c--628f8595ab45cb7cba7116fd__iPhone 12 PRO__iPhone X.MOV
-live_selfie/0001e96803--6239cf6720237d576aafbe45__Samsung Galaxy a22.mp4,live_video/0001e96803--6239cf6720237d576aafbe45__Samsung Galaxy a22.mp4,cut-out printouts/0001e96803--6239cf6720237d576aafbe45.mp4,printouts/0001e96803--6239cf6720237d576aafbe45.mp4,replay/0001ffba3c--629066494796d7421dad66b7__iPhone xr __iPhone 6s.MOV
-live_selfie/0001e96803--624508e961974a65884d9fd3__Redmi note 9.mp4,live_video/0001e96803--624508e961974a65884d9fd3__Redmi note 9.mp4,cut-out printouts/0001e96803--624508e961974a65884d9fd3.mp4,printouts/0001e96803--624508e961974a65884d9fd3.mp4,replay/0001ffba3c--6295ba1f8753764e91dcf5df__Poco X3 Pro__Tecno Pouvoir 4.mp4
-live_selfie/0001e96803--625473b64949293a58c89e41__Oppo a5s.mp4,live_video/0001e96803--625473b64949293a58c89e41__Oppo a5s.mp4,cut-out printouts/0001e96803--625473b64949293a58c89e41.mp4,printouts/0001e96803--625473b64949293a58c89e41.mp4,replay/0001ffba3c--62960f430bd50b2755f74916__Galaxy M31__M2003J15SG.mp4
