--- Anti_Spoofing_Project_using_DL/Model/convert.py
@@ -1,7 +1,0 @@
-from keras.models import load_model
-
-# Load the model
-# model = load_model('anti_spoofing_project_using_dl.ipynb')
-
-# Save the model in HDF5 format
-model.save('anti_spoofing_project_using_dl.h5')
