--- Anti_Spoofing_Project_using_DL/WebApp/templates/index.html
@@ -1,40 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Video Upload</title>
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css">
-    <style>
-        body {
-            background-image: url('https://www.pymnts.com/wp-content/uploads/2019/12/Biometric-Authentication-Digital-ID-Spoofing.jpg');
-            background-size: cover;
-            background-repeat: no-repeat;
-            background-position: center;
-            height: 100vh; /* Set background to cover the full viewport height */
-        }
-    </style>
-</head>
-<body>
-    <h1 style="background-color: aqua ;   text-align: center;"> Welcome To Anti-Spoofing System</h1>
-    <div class="container mt-5">
-        <div class="card">
-            <div class="card-body">
-                <h3 class="card-title"> Please Upload a Video</h3>
-                {% if error %}
-                    <div class="alert alert-danger" role="alert">
-                        {{ error }}
-                    </div>
-                {% endif %}
-                <form action="/upload" method="post" enctype="multipart/form-data">
-                    <div class="form-group">
-                        <label for="file">Choose file</label>
-                        <input type="file" class="form-control" id="file" name="file">
-                    </div>
-                    <button type="submit" class="btn btn-primary">Upload and Detect</button>
-                </form>
-            </div>
-        </div>
-    </div>
-</body>
-</html>
