--- Anti_Spoofing_Project_using_DL/requirements.txt
@@ -1,5 +1,0 @@
-opencv-python
-numpy
-scikit-learn
-tensorflow
-keras
